+++ ViewItem.js
@@ -17,7 +17,7 @@ import AppIcon from '@folio/stripes-components/lib/AppIcon';
 import craftLayerUrl from '@folio/stripes-components/util/craftLayerUrl';
 
 import ItemForm from './edit/items/ItemForm';
-import ViewMetaData from './ViewMetaData';
+import ViewMetadata from './ViewMetadata';
 
 class ViewItem extends React.Component {
   static manifest = Object.freeze({
@@ -69,7 +69,7 @@ class ViewItem extends React.Component {
       type: 'okapi',
       path: 'users?query=(id==%{borrowerId.query})',
       records: 'users',
-    }
+    },
   });
 
   /**
@@ -112,9 +112,6 @@ class ViewItem extends React.Component {
       // FIXME: loan-status-check must be i18n friendly
       if (nextProps.itemId === loan.itemId && loan.item.status.name !== 'Available') {
         nextProps.mutator.borrowerId.replace({ query: loan.userId });
-
-        // don't choke while we migrate from metaData to metadata
-        loan.metadata = loan.metadata || loan.metaData;
         return { loan };
       }
 
@@ -145,7 +142,7 @@ class ViewItem extends React.Component {
     };
 
     this.craftLayerUrl = craftLayerUrl.bind(this);
-    this.cViewMetaData = props.stripes.connect(ViewMetaData);
+    this.cViewMetadata = props.stripes.connect(ViewMetadata);
   }
 
   onClickEditItem = (e) => {
@@ -199,6 +196,7 @@ class ViewItem extends React.Component {
       referenceTables,
       okapi, stripes: { intl, formatDateTime } } = this.props;
 
+    const formatMsg = intl.formatMessage;
     referenceTables.shelfLocations = (shelfLocations || {}).records || [];
     referenceTables.loanTypes = (loanTypes || {}).records || [];
     referenceTables.materialTypes = (materialTypes || {}).records || [];
@@ -261,7 +259,7 @@ class ViewItem extends React.Component {
               <div style={{ textAlign: 'center' }}>
                 <AppIcon app="inventory" iconKey="item" size="small" /> {_.get(item, ['barcode'], '')}
                 <div>
-                  Item . {_.get(item, ['status', 'name'], '')}
+                  { formatMsg({ id: 'ui-inventory.itemDotStatus' }, { status: _.get(item, ['status', 'name'], '') }) }
                 </div>
               </div>
             }
@@ -271,12 +269,12 @@ class ViewItem extends React.Component {
           >
             <Row center="xs">
               <Col sm={6}>
-                Instance: {instance.title}
+                {formatMsg({ id: 'ui-inventory.instance' })} {instance.title}
                 {(instance.publication && instance.publication.length > 0) &&
                 <span><em>, </em><em>{instance.publication[0].publisher}{instance.publication[0].dateOfPublication ? `, ${instance.publication[0].dateOfPublication}` : ''}</em></span>
                 }
                 <div>
-                  {`Holdings: ${labelLocation} > ${labelCallNumber}`}
+                  { `${formatMsg({ id: 'ui-inventory.holdingsColon' })} ${labelLocation} > ${labelCallNumber}`}
                 </div>
               </Col>
             </Row>
@@ -284,16 +282,16 @@ class ViewItem extends React.Component {
               open={this.state.accordions.itemAccordion}
               id="itemAccordion"
               onToggle={this.handleAccordionToggle}
-              label="Item data"
+              label={formatMsg({ id: 'ui-inventory.itemData' })}
             >
               <Row>
                 <Col sm={12}>
-                  <AppIcon app="inventory" iconKey="item" size="small" /> Item record <AppIcon app="inventory" iconKey="material-type" size="small" /> {_.get(item, ['materialType', 'name'], '')} <AppIcon app="inventory" iconKey="item-status" size="small" /> {_.get(item, ['status', 'name'], '')}
+                  <AppIcon app="inventory" iconKey="item" size="small" /> {formatMsg({ id: 'ui-inventory.itemRecord' })} <AppIcon app="inventory" iconKey="material-type" size="small" /> {_.get(item, ['materialType', 'name'], '')} <AppIcon app="inventory" iconKey="item-status" size="small" /> {_.get(item, ['status', 'name'], '')}
                 </Col>
               </Row>
               <br />
               { (item.metadata && item.metadata.createdDate) &&
-                <this.cViewMetaData metadata={item.metadata} />
+                <this.cViewMetadata metadata={item.metadata} />
               }
               { (item.barcode) &&
                 <Row>
@@ -307,62 +305,62 @@ class ViewItem extends React.Component {
               <br /><br />
               <Row>
                 <Col xs={12}>
-                  <KeyValue label="Item ID" value={_.get(item, ['id'], '')} />
+                  <KeyValue label={formatMsg({ id: 'ui-inventory.itemId' })} value={_.get(item, ['id'], '')} />
                 </Col>
               </Row>
               { (item.temporaryLocation) &&
                 <Row>
                   <Col smOffset={0} sm={4}>
-                    <KeyValue label="Temporary location" value={_.get(item, ['temporaryLocation', 'name'], '')} />
+                    <KeyValue label={formatMsg({ id: 'ui-inventory.temporaryLocation' })} value={_.get(item, ['temporaryLocation', 'name'], '')} />
                   </Col>
                 </Row>
               }
               { (item.permanentLoanType) &&
                 <Row>
                   <Col smOffset={0} sm={4}>
-                    <KeyValue label="Permanent loantype" value={_.get(item, ['permanentLoanType', 'name'], '')} />
+                    <KeyValue label={formatMsg({ id: 'ui-inventory.permanentLoantype' })} value={_.get(item, ['permanentLoanType', 'name'], '')} />
                   </Col>
                 </Row>
               }
               { (item.temporaryLoanType) &&
                 <Row>
                   <Col smOffset={0} sm={4}>
-                    <KeyValue label="Temporary loantype" value={_.get(item, ['temporaryLoanType', 'name'], '')} />
+                    <KeyValue label={formatMsg({ id: 'ui-inventory.temporaryLoantype' })} value={_.get(item, ['temporaryLoanType', 'name'], '')} />
                   </Col>
                 </Row>
               }
               { (item.enumeration) &&
                 <Row>
                   <Col smOffset={0} sm={4}>
-                    <KeyValue label="Enumeration" value={_.get(item, ['enumeration'], '')} />
+                    <KeyValue label={formatMsg({ id: 'ui-inventory.enumeration' })} value={_.get(item, ['enumeration'], '')} />
                   </Col>
                 </Row>
               }
               { (item.chronology) &&
                 <Row>
                   <Col smOffset={0} sm={4}>
-                    <KeyValue label="Chronology" value={_.get(item, ['chronology'], '')} />
+                    <KeyValue label={formatMsg({ id: 'ui-inventory.chronology' })} value={_.get(item, ['chronology'], '')} />
                   </Col>
                 </Row>
               }
               { (item.numberOfPieces) &&
                 <Row>
                   <Col smOffset={0} sm={4}>
-                    <KeyValue label="Number of pieces" value={_.get(item, ['numberOfPieces'], '')} />
+                    <KeyValue label={formatMsg({ id: 'ui-inventory.numberOfPieces' })} value={_.get(item, ['numberOfPieces'], '')} />
                   </Col>
                 </Row>
               }
               { (item.pieceIdentifiers) &&
                 <Row>
                   <Col smOffset={0} sm={4}>
-                    <KeyValue label="Piece identifiers" value={_.get(item, ['pieceIdentifiers'], []).map((line, i) => <div key={i}>{line}</div>)} />
+                    <KeyValue label={formatMsg({ id: 'ui-inventory.pieceIdentifiers' })} value={_.get(item, ['pieceIdentifiers'], []).map((line, i) => <div key={i}>{line}</div>)} />
                   </Col>
                 </Row>
               }
               { (item.notes.length > 0) &&
                 <Row>
                   <Col smOffset={0} sm={4}>
-                    <KeyValue label="Notes" value={_.get(item, ['notes'], []).map((line, i) => <div key={i}>{line}</div>)} />
+                    <KeyValue label={formatMsg({ id: 'ui-inventory.notes' })} value={_.get(item, ['notes'], []).map((line, i) => <div key={i}>{line}</div>)} />
                   </Col>
                 </Row>
               }
@@ -408,6 +406,7 @@ class ViewItem extends React.Component {
             instance={instance}
             holdingsRecord={holdingsRecord}
             referenceTables={referenceTables}
+            intl={intl}
           />
         </Layer>
         <Layer isOpen={query.layer === 'copyItem'} label="Copy Item Dialog">
@@ -421,6 +420,7 @@ class ViewItem extends React.Component {
             copy
             holdingsRecord={holdingsRecord}
             referenceTables={referenceTables}
+            intl={intl}
           />
         </Layer>