+++ edit/InstanceForm.js
@@ -25,11 +25,12 @@ import DescriptionFields from './descriptionFields';
 import NoteFields from './noteFields';
 import LanguageFields from './languageFields';
 
-function validate(values) {
+function validate(values, props) {
   const errors = {};
+  const formatMsg = props.stripes.intl.formatMessage;
 
-  const requiredTextMessage = 'Please fill this in to continue';
-  const requiredSelectMessage = 'Please select to continue';
+  const requiredTextMessage = formatMsg({ id: 'ui-inventory.fillIn' });
+  const requiredSelectMessage = formatMsg({ id: 'ui-inventory.selectToContinue' });
 
   if (!values.title) {
     errors.title = requiredTextMessage;
@@ -114,6 +115,7 @@ class InstanceForm extends React.Component {
       referenceTables,
       copy,
     } = this.props;
+    const formatMsg = this.props.stripes.intl.formatMessage;
 
     const instanceTypeOptions = referenceTables.instanceTypes ? referenceTables.instanceTypes.map(
       it => ({
@@ -132,22 +134,22 @@ class InstanceForm extends React.Component {
     ) : [];
 
     /* Menus for Add Instance workflow */
-    const addInstanceLastMenu = <PaneMenu><Button buttonStyle="primary paneHeaderNewButton" id="clickable-create-instance" type="submit" title="Create New Instance" disabled={(pristine || submitting) && !copy} onClick={handleSubmit}>Create instance</Button></PaneMenu>;
-    const editInstanceLastMenu = <PaneMenu><Button buttonStyle="primary paneHeaderNewButton" id="clickable-update-instance" type="submit" title="Update Instance" disabled={(pristine || submitting) && !copy} onClick={handleSubmit}>Update instance</Button></PaneMenu>;
+    const addInstanceLastMenu = <PaneMenu><Button buttonStyle="primary paneHeaderNewButton" id="clickable-create-instance" type="submit" title={formatMsg({ id: 'ui-inventory.createInstance' })} disabled={(pristine || submitting) && !copy} onClick={handleSubmit}>{formatMsg({ id: 'ui-inventory.createInstance' })}</Button></PaneMenu>;
+    const editInstanceLastMenu = <PaneMenu><Button buttonStyle="primary paneHeaderNewButton" id="clickable-update-instance" type="submit" title={formatMsg({ id: 'ui-inventory.updateInstance' })} disabled={(pristine || submitting) && !copy} onClick={handleSubmit}>{formatMsg({ id: 'ui-inventory.updateInstance' })}</Button></PaneMenu>;
     return (
       <form>
         <Paneset isRoot>
           <Pane defaultWidth="100%" dismissible onClose={onCancel} lastMenu={initialValues.id ? editInstanceLastMenu : addInstanceLastMenu} paneTitle={initialValues.id ? 'Edit Instance' : 'New Instance'}>
             <Row>
-              <Col sm={12}><Headline size="large" tag="h3">Instance Record</Headline>
+              <Col sm={12}><Headline size="large" tag="h3">{formatMsg({ id: 'ui-inventory.instanceRecord' })}</Headline>
 
-                <Accordion label={<h3>Title data</h3>} onToggle={this.onToggleSection} open={this.state.sections.instanceSection1} id="instanceSection1">
+                <Accordion label={<h3>{formatMsg({ id: 'ui-inventory.titleData' })}</h3>} onToggle={this.onToggleSection} open={this.state.sections.instanceSection1} id="instanceSection1">
                   <Row>
                     <Col sm={9}>
                       <Row>
                         <Col sm={8}>
                           <Field
-                            label="Title *"
+                            label={formatMsg({ id: 'ui-inventory.title' })}
                             name="title"
                             id="input_instance_title"
                             component={TextField}
@@ -165,7 +167,7 @@ class InstanceForm extends React.Component {
                           />
                         </Col>
                       </Row>
-                      <AlternativeTitles />
+                      <AlternativeTitles formatMsg={formatMsg} />
                       <ContributorFields contributorNameTypes={referenceTables.contributorNameTypes} contributorTypes={referenceTables.contributorTypes} />
                       <Row>
                         <Col sm={8}>
@@ -174,8 +176,8 @@ class InstanceForm extends React.Component {
                             id="select_instance_type"
                             type="text"
                             component={Select}
-                            label="Resource type *"
-                            dataOptions={[{ label: 'Select resource type', value: '' }, ...instanceTypeOptions]}
+                            label={formatMsg({ id: 'ui-inventory.resourceType' })}
+                            dataOptions={[{ label: formatMsg({ id: 'ui-inventory.selectResourceType' }), value: '' }, ...instanceTypeOptions]}
                             required
                           />
                         </Col>
@@ -186,37 +188,37 @@ class InstanceForm extends React.Component {
                             name="instanceFormatId"
                             type="text"
                             component={Select}
-                            label="Format"
-                            dataOptions={[{ label: 'Select format', value: '' }, ...instanceFormatOptions]}
+                            label={formatMsg({ id: 'ui-inventory.format' })}
+                            dataOptions={[{ label: formatMsg({ id: 'ui-inventory.selectFormat' }), value: '' }, ...instanceFormatOptions]}
                           />
                         </Col>
                       </Row>
-                      <IdentifierFields identifierTypes={referenceTables.identifierTypes} />
+                      <IdentifierFields identifierTypes={referenceTables.identifierTypes} formatMsg={formatMsg} />
                     </Col>
                   </Row>
                 </Accordion>
-                <Accordion label={<h3>Descriptive data</h3>} onToggle={this.onToggleSection} open={this.state.sections.instanceSection2} id="instanceSection2">
+                <Accordion label={<h3>{formatMsg({ id: 'ui-inventory.descriptiveData' })}</h3>} onToggle={this.onToggleSection} open={this.state.sections.instanceSection2} id="instanceSection2">
                   <Row>
                     <Col sm={9}>
-                      <PublicationFields />
+                      <PublicationFields formatMsg={formatMsg} />
                       <Row>
                         <Col sm={8}>
-                          <Field label="Edition" name="edition" id="input_instance_edition" component={TextField} fullWidth />
+                          <Field label={formatMsg({ id: 'ui-inventory.edition' })} name="edition" id="input_instance_edition" component={TextField} fullWidth />
                         </Col>
                       </Row>
-                      <DescriptionFields />
-                      <LanguageFields />
-                      <URLFields />
+                      <DescriptionFields formatMsg={formatMsg} />
+                      <LanguageFields formatMsg={formatMsg} />
+                      <URLFields formatMsg={formatMsg} />
                     </Col>
                   </Row>
                 </Accordion>
-                <Accordion label={<h3>Content data</h3>} onToggle={this.onToggleSection} open={this.state.sections.instanceSection3} id="instanceSection3">
+                <Accordion label={<h3>{formatMsg({ id: 'ui-inventory.contentData' })}</h3>} onToggle={this.onToggleSection} open={this.state.sections.instanceSection3} id="instanceSection3">
                   <Row>
                     <Col sm={9}>
-                      <SeriesFields />
-                      <SubjectFields />
-                      <ClassificationFields classificationTypes={referenceTables.classificationTypes} />
-                      <NoteFields />
+                      <SeriesFields formatMsg={formatMsg} />
+                      <SubjectFields formatMsg={formatMsg} />
+                      <ClassificationFields classificationTypes={referenceTables.classificationTypes} formatMsg={formatMsg} />
+                      <NoteFields formatMsg={formatMsg} />
                     </Col>
                   </Row>
                 </Accordion>
@@ -239,6 +241,7 @@ InstanceForm.propTypes = {
   initialValues: PropTypes.object,
   referenceTables: PropTypes.object.isRequired,
   copy: PropTypes.bool,
+  stripes: { intl: { formatMsg: PropTypes.func } },
 };
 
 export default stripesForm({