+++ edit/items/ItemForm.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { FormattedMessage } from 'react-intl';
 import { Field, FieldArray, SubmissionError } from 'redux-form';
 import Paneset from '@folio/stripes-components/lib/Paneset';
 import Pane from '@folio/stripes-components/lib/Pane';
@@ -11,6 +12,7 @@ import TextField from '@folio/stripes-components/lib/TextField';
 import stripesForm from '@folio/stripes-form';
 import LocationSelection from '@folio/stripes-smart-components/lib/LocationSelection';
 import LocationLookup from '@folio/stripes-smart-components/lib/LocationLookup';
+import ConfirmationModal from '@folio/stripes-components/lib/structures/ConfirmationModal';
 
 import renderNotes from './noteFields';
 import renderPieceIdentifiers from './pieceIdentifierFields';
@@ -18,13 +20,14 @@ import renderPieceIdentifiers from './pieceIdentifierFields';
 
 function validate(values) {
   const errors = {};
+  const selectToContinueMsg = <FormattedMessage id="ui-inventory.selectToContinue" />;
 
   if (!(values.materialType && values.materialType.id)) {
-    errors.materialType = { id: 'Please select to continue' };
+    errors.materialType = { id: selectToContinueMsg };
   }
 
   if (!(values.permanentLoanType && values.permanentLoanType.id)) {
-    errors.permanentLoanType = { id: 'Please select to continue' };
+    errors.permanentLoanType = { id: selectToContinueMsg };
   }
 
   return errors;
@@ -38,6 +41,8 @@ function checkUniqueBarcode(okapi, barcode) {
 }
 
 function asyncValidate(values, dispatch, props, blurredField) {
+  const barcodeTakenMsg = props.intl.formatMessage({ id: 'ui-inventory.barcodeTaken' });
+
   if (blurredField === 'barcode' && values.barcode !== props.initialValues.barcode) {
     return new Promise((resolve, reject) => {
       // TODO: Should use stripes-connect (dispatching an action and update state)
@@ -47,7 +52,7 @@ function asyncValidate(values, dispatch, props, blurredField) {
         } else {
           response.json().then((json) => {
             if (json.totalRecords > 0) {
-              const error = new SubmissionError({ barcode: 'This barcode has already been taken' });
+              const error = new SubmissionError({ barcode: barcodeTakenMsg });
               reject(error);
             } else {
               resolve();
@@ -61,8 +66,39 @@ function asyncValidate(values, dispatch, props, blurredField) {
 }
 
 class ItemForm extends React.Component {
+  constructor() {
+    super();
+    this.state = {
+      confirmLocation: false,
+    };
+  }
+
+  componentDidMount() {
+    const { initialValues } = this.props;
+    const prevLocation = initialValues.temporaryLocation || {};
+    // eslint-disable-next-line react/no-did-mount-set-state
+    this.setState({ prevLocation });
+  }
+
   selectLocation(location) {
-    this.props.change('temporaryLocation.id', location.id);
+    if (!location.id) return;
+
+    if (location.isActive) {
+      setTimeout(() => this.props.change('temporaryLocation.id', location.id));
+      this.setState({ prevLocation: location });
+    } else {
+      this.setState({ confirmLocation: true, location });
+    }
+  }
+
+  confirmLocation(confirm) {
+    const { location, prevLocation } = this.state;
+    const confirmLocation = false;
+    const value = (confirm) ? location.id : prevLocation.id;
+    const prevLoc = (confirm) ? location : prevLocation;
+
+    setTimeout(() => this.props.change('temporaryLocation.id', value));
+    this.setState({ confirmLocation, prevLocation: prevLoc });
   }
 
   render() {
@@ -78,6 +114,9 @@ class ItemForm extends React.Component {
       referenceTables,
       copy,
     } = this.props;
+    const formatMsg = this.props.intl.formatMessage;
+
+    const { confirmLocation } = this.state;
 
     /* Menus for Add Item workflow */
     const addItemLastMenu = <PaneMenu><Button buttonStyle="primary paneHeaderNewButton" id="clickable-create-item" type="submit" title="Create New Item" disabled={(pristine || submitting) && !copy} onClick={handleSubmit}>Create item</Button></PaneMenu>;
@@ -124,66 +163,69 @@ class ItemForm extends React.Component {
           >
             <Row>
               <Col sm={5} smOffset={1}>
-                <h2>Item Record</h2>
+                <h2>
+                  <FormattedMessage id="ui-inventory.itemRecord" />
+                </h2>
               </Col>
             </Row>
             <Row >
               <Col sm={5} smOffset={1}>
                 {/* <Field label="Material Type" name="materialType.name" id="additem_materialType" component={TextField} fullWidth /> */}
                 <Field
-                  label="Material Type *"
+                  label={formatMsg({ id: 'ui-inventory.materialType' })}
                   name="materialType.id"
                   id="additem_materialType"
                   component={Select}
                   fullWidth
-                  dataOptions={[{ label: 'Select material type', value: '' }, ...materialTypeOptions]}
+                  dataOptions={[{ label: formatMsg({ id: 'ui-inventory.selectMaterialType' }), value: '' }, ...materialTypeOptions]}
                 />
                 <Field
-                  label="Loan Type (Permanent) *"
+                  label={formatMsg({ id: 'ui-inventory.loanTypePermanent' })}
                   name="permanentLoanType.id"
                   id="additem_loanTypePerm"
                   component={Select}
                   fullWidth
-                  dataOptions={[{ label: 'Select loan type', value: '' }, ...loanTypeOptions]}
+                  dataOptions={[{ label: formatMsg({ id: 'ui-inventory.selectLoanType' }), value: '' }, ...loanTypeOptions]}
                 />
-                <Field label="Barcode" name="barcode" id="additem_barcode" component={TextField} required fullWidth />
+                <Field label={formatMsg({ id: 'ui-inventory.barcode' })} name="barcode" id="additem_barcode" component={TextField} required fullWidth />
 
                 <Field
-                  label="Temporary Location"
-                  placeholder="Select temporary location"
+                  label={formatMsg({ id: 'ui-inventory.temporaryLocation' })}
+                  placeholder={formatMsg({ id: 'ui-inventory.selectTemporaryLocation' })}
                   name="temporaryLocation.id"
                   id="additem_location"
                   component={LocationSelection}
                   fullWidth
                   marginBottom0
+                  onChange={loc => this.selectLocation(loc)}
                 />
                 <LocationLookup temporary onLocationSelected={loc => this.selectLocation(loc)} />
 
-                <Field label="Status" name="status.name" id="additem_status" component={TextField} disabled fullWidth />
+                <Field label={formatMsg({ id: 'ui-inventory.status' })} name="status.name" id="additem_status" component={TextField} disabled fullWidth />
                 <Field
-                  label="Loan Type (Temporary)"
+                  label={formatMsg({ id: 'ui-inventory.loanTypeTemporary' })}
                   name="temporaryLoanType.id"
                   id="additem_loanTypeTemp"
                   component={Select}
                   fullWidth
-                  dataOptions={[{ label: 'Select loan type', value: '' }, ...loanTypeOptions]}
+                  dataOptions={[{ label: formatMsg({ id: 'ui-inventory.selectLoanType' }), value: '' }, ...loanTypeOptions]}
                 />
                 <Field
-                  label="Enumeration"
+                  label={formatMsg({ id: 'ui-inventory.enumeration' })}
                   name="enumeration"
                   id="additem_enumeration"
                   component={TextField}
                   fullWidth
                 />
                 <Field
-                  label="Chronology"
+                  label={formatMsg({ id: 'ui-inventory.chronology' })}
                   name="chronology"
                   id="additem_chronology"
                   component={TextField}
                   fullWidth
                 />
                 <Field
-                  label="Number of pieces"
+                  label={formatMsg({ id: 'ui-inventory.numberOfPieces' })}
                   name="numberOfPieces"
                   id="additem_numberofpieces"
                   component={TextField}
@@ -192,15 +234,23 @@ class ItemForm extends React.Component {
             </Row>
             <Row>
               <Col sm={8} smOffset={1}>
-                <FieldArray name="notes" component={renderNotes} />
+                <FieldArray name="notes" component={renderNotes} formatMsg={formatMsg} />
               </Col>
             </Row>
             <Row>
               <Col sm={8} smOffset={1}>
-                <FieldArray name="pieceIdentifiers" component={renderPieceIdentifiers} />
+                <FieldArray name="pieceIdentifiers" component={renderPieceIdentifiers} formatMsg={formatMsg} />
               </Col>
             </Row>
 
+            <ConfirmationModal
+              open={confirmLocation}
+              heading={formatMsg({ id: 'ui-inventory.confirmLocation.header' })}
+              message={formatMsg({ id: 'ui-inventory.confirmLocation.message' })}
+              confirmLabel={formatMsg({ id: 'ui-inventory.confirmLocation.selectBtn' })}
+              onConfirm={() => { this.confirmLocation(true); }}
+              onCancel={() => { this.confirmLocation(false); }}
+            />
           </Pane>
         </Paneset>
       </form>
@@ -222,6 +272,9 @@ ItemForm.propTypes = {
   holdingsRecord: PropTypes.object,
   referenceTables: PropTypes.object.isRequired,
   copy: PropTypes.bool,
+  intl: PropTypes.shape({
+    formatMessage: PropTypes.func,
+  }).isRequired,
 };
 
 export default stripesForm({