+++ edit/items/noteFields.js
@@ -5,11 +5,11 @@ import { Row, Col } from '@folio/stripes-components/lib/LayoutGrid';
 import Button from '@folio/stripes-components/lib/Button';
 import TextField from '@folio/stripes-components/lib/TextField';
 
-const renderNotes = ({ fields, meta: { touched, error, submitFailed } }) => (
+const renderNotes = ({ fields, meta: { touched, error, submitFailed }, formatMsg }) => (
   <div>
     <Row>
       <Col sm={2} smOffset={5}>
-        <Button type="button" buttonStyle="fullWidth secondary" id="clickable-add-notes" onClick={() => fields.push()}>Add Notes</Button>
+        <Button type="button" buttonStyle="fullWidth secondary" id="clickable-add-notes" onClick={() => fields.push()}>{formatMsg({ id: 'ui-inventory.addNotes' })}</Button>
         {(touched || submitFailed) && error && <span>{error}</span>}
       </Col>
     </Row>
@@ -20,7 +20,7 @@ const renderNotes = ({ fields, meta: { touched, error, submitFailed } }) => (
             name={note}
             type="text"
             component={TextField}
-            label={index === 0 ? 'Notes' : null}
+            label={index === 0 ? formatMsg({ id: 'ui-inventory.notes' }) : null}
           />
         </Col>
         <Col sm={1}>
@@ -28,14 +28,14 @@ const renderNotes = ({ fields, meta: { touched, error, submitFailed } }) => (
           <Button
             buttonStyle="fullWidth secondary"
             type="button"
-            title={`Remove Notes ${index + 1}`}
+            title={formatMsg({ id: 'ui-inventory.removeNotes' }, { num: index + 1 })}
             onClick={() => fields.remove(index)}
-          >Remove
+          >{formatMsg({ id: 'ui-inventory.remove' })}
           </Button>
         </Col>
       </Row>)}
   </div>
 );
-renderNotes.propTypes = { fields: PropTypes.object, meta: PropTypes.object };
+renderNotes.propTypes = { fields: PropTypes.object, meta: PropTypes.object, formatMsg: PropTypes.func };
 
 export default renderNotes;