+++ edit/languageFields.js
@@ -5,14 +5,14 @@ import Select from '@folio/stripes-components/lib/Select';
 import RepeatableField from '@folio/stripes-components/lib/structures/RepeatableField';
 import languages from '../data/languages';
 
-const renderLanguageField = ({ field, fieldIndex }) => {
+const renderLanguageField = ({ field, fieldIndex }, formatMsg) => {
   const languageOptions = languages.selectOptions(field);
   return (
     <Field
-      label={fieldIndex === 0 ? 'language' : null}
+      label={fieldIndex === 0 ? formatMsg({ id: 'ui-inventory.language' }) : null}
       name={`${field}`}
       component={Select}
-      dataOptions={[{ label: 'Select language', value: '' }, ...languageOptions]}
+      dataOptions={[{ label: formatMsg({ id: 'ui-inventory.selectLanguage' }), value: '' }, ...languageOptions]}
     />
   );
 };
@@ -22,16 +22,17 @@ renderLanguageField.propTypes = {
   fieldIndex: PropTypes.number,
 };
 
-const LanguageFields = () => (
+const LanguageFields = ({ formatMsg }) => (
   <RepeatableField
     name="languages"
-    label="Languages"
-    addLabel="+ Add language"
+    label={formatMsg({ id: 'ui-inventory.languages' })}
+    addLabel={formatMsg({ id: 'ui-inventory.addLanguage' })}
     addButtonId="clickable-add-language"
     template={[{
-      render: renderLanguageField,
-    }]}
+        render(fieldObj) { return renderLanguageField(fieldObj, formatMsg); },
+      }]}
   />
 );
+LanguageFields.propTypes = { formatMsg: PropTypes.func };
 
 export default LanguageFields;