+++ CHANGELOG.md
@@ -63,6 +63,11 @@
 * Migrate to hierarchical location end-point. Fixes UIIN-127.
 * Handle "metaData" field case insensitively for compatibility. Fixes UIIN-159.
 * Only show item-availabilty for open loans. Refs UIIN-110.
+* Add location-units and locations to Okapi interfaces. Fixes UIIN-171.
+* Correct use of i18n in validation function and props. Fixes UIIN-172.
+* Confirmation when assigning inactive location. Fixes UIIN-121.
+* Bah; metadata. One case to rule them all. Refs UIIN-159, CIRCSTORE-43.
+* Fix holding and item forms. Fixes UIIN-174.
 
 ## [1.0.0](https://github.com/folio-org/ui-instances/tree/v1.0.0) (2017-09-08)
 [Full Changelog](https://github.com/folio-org/ui-items/compare/v0.0.1...v1.0.0)